requirements.txt
gpflow == 2.2.1
hyperopt == 0.2.7
joblib == 1.1.0
matplotlib == 3.5.0
mysql_connector_python == 8.0.28
numpy == 1.19.5
ordered_set == 4.0.2
pandas == 1.3.4
scikit_learn == 1.0.2
seaborn == 0.11.2
tensorflow == 2.5.3
tensorflow_probability == 0.13.0
trieste == 0.10.0
xgboost == 1.5.2